# openapi.yaml
openapi: 3.0.1
info:
  title: Kramesh Portfolio Backend API
  description: API for managing portfolio content (education, experience, skills, about, projects, blogs, contact, certifications, auth).
  version: "1.0"
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /contact:
    post:
      tags:
        - contact
      operationId: sendMessage
      summary: Send a contact message
      description: Public endpoint â€” visitors submit the contact form. No auth required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactFields"
      responses:
        201:
          description: Message sent successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Message sent successfully!
                required:
                  - message
        400:
          description: Validation failed (missing name, email, or message).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_fields:
                  value:
                    message: Name, email and message are required
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/contact_post_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /contact/messages:
    get:
      tags:
        - contact
      operationId: getMessages
      summary: List contact messages (admin)
      description: Returns all submitted contact messages. Requires admin authentication.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Messages found and returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/contact_messages_get_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/contact_messages_get_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /contact/messages/{id}:
    delete:
      tags:
        - contact
      operationId: deleteMessage
      summary: Delete a contact message by ID
      description: Deletes a contact message. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d6
      responses:
        200:
          description: Message deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMessage"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/contact_messages_delete_unauthorized.json
  /certifications:
    get:
      tags:
        - certifications
      operationId: getCertifications
      summary: List all certifications
      description: Returns an array of certification records, sorted by display_order.
      responses:
        200:
          description: Certifications found and returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certification"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/certifications_get_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
    post:
      tags:
        - certifications
      operationId: createCertification
      summary: Create a certification
      description: Creates a new certification record. Requires admin authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationFields"
      responses:
        201:
          description: Certification created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/certifications_post_unauthorized.json
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/certifications_post_server_error.json
      parameters:
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
  /certifications/{id}:
    put:
      tags:
        - certifications
      operationId: updateCertification
      summary: Update a certification by ID
      description: Updates an existing certification record. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d7
        - name: X-Simulate-Error
          in: header
          required: false
          schema:
            type: string
          description: Test-only header (NODE_ENV=test) to simulate a 500 server error for contract testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificationFields"
      responses:
        200:
          description: Certification updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
        400:
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/certifications_put_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                server_error:
                  $ref: openapi_examples/certifications_put_server_error.json
    delete:
      tags:
        - certifications
      operationId: deleteCertification
      summary: Delete a certification by ID
      description: Deletes a certification record. Requires admin authentication.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          example: 64f1a2b3c4d5e6f7a8b9c0d7
      responses:
        200:
          description: Certification deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMessage"
        401:
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthorized:
                  $ref: openapi_examples/certifications_delete_unauthorized.json
        404:
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      title: Error response
      type: object
      properties:
        message:
          type: string
          example: Not authorized
      required:
        - message
    DeleteMessage:
      title: Delete confirmation message
      type: object
      properties:
        message:
          type: string
          example: Education deleted
      required:
        - message
...
    ContactFields:
      title: Contact fields
      description: Fields submitted through the contact form.
      type: object
      properties:
        name:
          type: string
          example: Priya Sharma
        email:
          type: string
          example: priya@example.com
        subject:
          type: string
          example: Project collaboration inquiry
        message:
          type: string
          example: Hi, I'd like to discuss a potential collaboration.
      required:
        - name
        - email
        - message
    Contact:
      title: Contact message
      description: A stored contact-form submission.
      allOf:
        - $ref: "#/components/schemas/ContactFields"
        - type: object
          properties:
            _id:
              type: string
              pattern: ^[0-9a-fA-F]{24}$
              example: 64f1a2b3c4d5e6f7a8b9c0d6
            is_read:
              type: boolean
              example: false
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
              example: 0
          required:
            - _id
    CertificationFields:
      title: Certification fields
      description: Editable fields of a certification.
      type: object
      properties:
        institution_name:
          type: string
          example: IIT Roorkee
        certificate_name:
          type: string
          example: Executive PG Certification in Full Stack Web Development
        certificate_id:
          type: string
          example: IITR-FSD-2026-001
        description:
          type: string
          example: A comprehensive program covering MERN stack and Java/Spring Boot.
        skills:
          type: string
          example: React, Node.js, Java, Spring Boot
        start_date:
          type: string
          example: "2025-06-01"
        end_date:
          type: string
          example: "2026-05-30"
        is_expired:
          type: boolean
          example: false
        credential_url:
          type: string
          example: https://example.com/verify/abc123
        image_url:
          type: string
          example: https://res.cloudinary.com/demo/image/upload/cert.jpg
        display_order:
          type: integer
          example: 1
      required:
        - institution_name
        - certificate_name
    Certification:
      title: Certification
      description: A certification entry.
      allOf:
        - $ref: "#/components/schemas/CertificationFields"
        - type: object
          properties:
            _id:
              type: string
              pattern: ^[0-9a-fA-F]{24}$
              example: 64f1a2b3c4d5e6f7a8b9c0d7
            image_public_id:
              type: string
              example: cert_abc123
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            __v:
              type: integer
              example: 0
          required:
            - _id
...